get_max_area_bf {mmtable2}	R Documentation

An algorthm to identify the largest rectangle in a histogram

Description

An algorthm to identify the largest rectangle in a histogram

Usage

get_max_area_bf(heights)

Arguments

heights
a list of the heights of the histogram.

Value

list

[Package mmtable2 version 0.1.3 Index]