single_cells_to_vector {mmtable2}	R Documentation

Convert a list of cells to a list of cell ranges

Description

Convert a list of cells to a list of cell ranges

Usage

single_cells_to_vector(row_col_list)

Arguments

row_col_list
list of cells

[Package mmtable2 version 0.1.3 Index]