table_constructor {mmtable2}	R Documentation

Use attributes to construct a table with column and row headers.

Description

Use attributes to construct a table with column and row headers.

Usage

table_constructor(
  df,
  col_header_df,
  row_header_df,
  data_vars,
  table_name = NULL,
  table_format
)

Arguments

df
a data frame

col_header_df
col header information

row_header_df
row header information

data_vars
cell variable

table_name
name of the table

table_format
formatting o the table

Value

mmtable

[Package mmtable2 version 0.1.3 Index]